+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## FIXME
+
+* BREAKING CHANGE: Started using **async/await**, so Node.JS >=7.6.0 is required.
+* FEATURE: Added Topic Report API - https://github.com/citizenos/citizenos-api/issues/5
+* Default Node.JS version upgraded to 10.13.0.
+
 ## 2019-04-10
 
 * Fix etherpad sync endpoint, remove rev number to put less stress on etherpad. * Fix partner topic read test